Remove method `isClosed` in interface `KvBackend`

+++ core/src/main/java/com/datastrato/gravitino/storage/kv/RocksDBKvBackend.java
@@ -75,6 +75,11 @@ public class RocksDBKvBackend implements KvBackend {
 
   @Override
   public void put(byte[] key, byte[] value, boolean overwrite) throws IOException {
+    if (hasClosed) {
+      throw new IllegalStateException(
+          "RocksDBKvBackend has been closed, put operation can't be perform...");
+    }
+
     try {
       handlePutWithoutTransaction(key, value, overwrite);
     } catch (EntityAlreadyExistsException e) {
@@ -102,6 +107,11 @@ public class RocksDBKvBackend implements KvBackend {
 
   @Override
   public byte[] get(byte[] key) throws IOException {
+    if (hasClosed) {
+      throw new IllegalStateException(
+          "RocksDBKvBackend has been closed, get operation can't be perform...");
+    }
+
     try {
       return db.get(key);
     } catch (RocksDBException e) {
@@ -111,6 +121,11 @@ public class RocksDBKvBackend implements KvBackend {
 
   @Override
   public List<Pair<byte[], byte[]>> scan(KvRangeScan scanRange) throws IOException {
+    if (hasClosed) {
+      throw new IllegalStateException(
+          "RocksDBKvBackend has been closed, scan operation can't be perform...");
+    }
+
     RocksIterator rocksIterator = db.newIterator();
     rocksIterator.seek(scanRange.getStart());
 
@@ -153,6 +168,11 @@ public class RocksDBKvBackend implements KvBackend {
 
   @Override
   public boolean delete(byte[] key) throws IOException {
+    if (hasClosed) {
+      throw new IllegalStateException(
+          "RocksDBKvBackend has been closed, delete operation can't be perform...");
+    }
+
     try {
       db.delete(key);
       return true;
@@ -163,6 +183,11 @@ public class RocksDBKvBackend implements KvBackend {
 
   @Override
   public boolean deleteRange(KvRangeScan deleteRange) throws IOException {
+    if (hasClosed) {
+      throw new IllegalStateException(
+          "RocksDBKvBackend has been closed, deleteRange operation can't be perform...");
+    }
+
     RocksIterator rocksIterator = db.newIterator();
     rocksIterator.seek(deleteRange.getStart());
 
@@ -197,10 +222,5 @@ public class RocksDBKvBackend implements KvBackend {
   public void close() throws IOException {
     db.close();
     hasClosed = true;
-  }
-
-  @Override
-  public synchronized boolean isClosed() {
-    return hasClosed;
   }
 }
